requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
Flask-Login==0.6.1
mysql-connector-python==8.0.33
PyMySQL==1.0.2
face_recognition==1.3.0
opencv-python==4.7.0.72
Flask-SocketIO==5.3.0
APScheduler==3.10.0
Flask-Migrate==4.0.4


pip install Flask
pip install Flask-SQLAlchemy
pip install Flask-Login
pip install mysql-connector-python
pip install PyMySQL==1.0.2
pip install face_recognition
pip install opencv-python
pip install Flask-S
